store: Move into internal
