+++ cmd/cometbft/commands/inspect.go
@@ -12,7 +12,7 @@ import (
 	"github.com/cometbft/cometbft/internal/inspect"
 	"github.com/cometbft/cometbft/internal/state"
 	"github.com/cometbft/cometbft/internal/state/indexer/block"
-	"github.com/cometbft/cometbft/store"
+	"github.com/cometbft/cometbft/internal/store"
 	"github.com/cometbft/cometbft/types"
 )
