+++ internal/consensus/byzantine_test.go
@@ -24,10 +24,10 @@ import (
 	"github.com/cometbft/cometbft/proxy"
 
 	sm "github.com/cometbft/cometbft/internal/state"
+	"github.com/cometbft/cometbft/internal/store"
 	"github.com/cometbft/cometbft/p2p"
 	cmtcons "github.com/cometbft/cometbft/proto/tendermint/consensus"
 	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
-	"github.com/cometbft/cometbft/store"
 	"github.com/cometbft/cometbft/types"
 )
