+++ internal/evidence/pool_test.go
@@ -15,10 +15,10 @@ import (
 	"github.com/cometbft/cometbft/internal/evidence/mocks"
 	sm "github.com/cometbft/cometbft/internal/state"
 	smmocks "github.com/cometbft/cometbft/internal/state/mocks"
+	"github.com/cometbft/cometbft/internal/store"
 	"github.com/cometbft/cometbft/internal/test"
 	"github.com/cometbft/cometbft/libs/log"
 	cmtversion "github.com/cometbft/cometbft/proto/tendermint/version"
-	"github.com/cometbft/cometbft/store"
 	"github.com/cometbft/cometbft/types"
 	"github.com/cometbft/cometbft/version"
 )
