+++ rpc/grpc/server/server.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	sm "github.com/cometbft/cometbft/internal/state"
-	"github.com/cometbft/cometbft/store"
+	"github.com/cometbft/cometbft/internal/store"
 
 	brs "github.com/cometbft/cometbft/proto/tendermint/services/block_results/v1"
 	"github.com/cometbft/cometbft/rpc/grpc/server/services/blockresultservice"
