+++ rpc/grpc/server/services/blockservice/service.go
@@ -8,11 +8,11 @@ import (
 	"google.golang.org/grpc/status"
 
 	"github.com/cometbft/cometbft/internal/rpctrace"
+	"github.com/cometbft/cometbft/internal/store"
 	"github.com/cometbft/cometbft/libs/log"
 	cmtpubsub "github.com/cometbft/cometbft/libs/pubsub"
 	blocksvc "github.com/cometbft/cometbft/proto/tendermint/services/block/v1"
 	ptypes "github.com/cometbft/cometbft/proto/tendermint/types"
-	"github.com/cometbft/cometbft/store"
 	"github.com/cometbft/cometbft/types"
 )
