+++ test/loadtime/cmd/report/main.go
@@ -11,7 +11,7 @@ import (
 
 	dbm "github.com/cometbft/cometbft-db"
 
-	"github.com/cometbft/cometbft/store"
+	"github.com/cometbft/cometbft/internal/store"
 	"github.com/cometbft/cometbft/test/loadtime/report"
 )
